refactor: 레거시 코드 삭제
