+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -16,9 +16,8 @@ public class AppWebApplicationInitializer implements WebApplicationInitializer {
     @Override
     public void onStartup(ServletContext servletContext) {
         final DispatcherServlet dispatcherServlet = new DispatcherServlet();
-        dispatcherServlet.addHandlerMapping(new AnnotationHandlerMapping("com.techcourse.controller.annotaion"),
-                new ManualHandlerMapping());
-        dispatcherServlet.addHandlerAdapter(new AnnotationHandlerAdapter(), new ManualHandlerAdapter());
+        dispatcherServlet.addHandlerMapping(new AnnotationHandlerMapping("com.techcourse.controller"));
+        dispatcherServlet.addHandlerAdapter(new AnnotationHandlerAdapter());
 
         final ServletRegistration.Dynamic dispatcher = servletContext.addServlet("dispatcher", dispatcherServlet);
         dispatcher.setLoadOnStartup(1);
